+++ template/src/app/(payload)/admin/importMap.js
@@ -1,10 +1,14 @@
-import { ImageButtons as ImageButtons_852fda9a9517933518bec4b451eab31a } from 'sitewright-core/ui'
+import { ImageButtons as ImageButtons_742e89665ed7ae7fb94e993adb58357d } from 'sitewright-core/ui'
+import { default as default_ce5b1c2f3c90e2cf6040e4bba58f7a23 } from '@/components/admin/GuiaLink'
 import { VercelBlobClientUploadHandler as VercelBlobClientUploadHandler_16c82c5e25f430251a3e3ba57219ff4e } from '@payloadcms/storage-vercel-blob/client'
+import { default as default_225f9126e2d95819f0e7fbaf3654ab1e } from '@/components/admin/Guia'
 import { CollectionCards as CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1 } from '@payloadcms/next/rsc'
 
 /** @type import('payload').ImportMap */
 export const importMap = {
-  "sitewright-core/ui#ImageButtons": ImageButtons_852fda9a9517933518bec4b451eab31a,
+  "sitewright-core/ui#ImageButtons": ImageButtons_742e89665ed7ae7fb94e993adb58357d,
+  "@/components/admin/GuiaLink#default": default_ce5b1c2f3c90e2cf6040e4bba58f7a23,
   "@payloadcms/storage-vercel-blob/client#VercelBlobClientUploadHandler": VercelBlobClientUploadHandler_16c82c5e25f430251a3e3ba57219ff4e,
+  "@/components/admin/Guia#default": default_225f9126e2d95819f0e7fbaf3654ab1e,
   "@payloadcms/next/rsc#CollectionCards": CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1
 }